Modify SettingPage split subpages into individual files
